% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdmmMC.R
\name{AdmmMC}
\alias{AdmmMC}
\title{MC6068}
\usage{
AdmmMC(x, iniMatrix = NULL, error = 1e-16, lambda = 0.5,
  silence = F)
}
\arguments{
\item{x}{Incomplete matrix of your interest}

\item{iniMatrix}{Matrix with initial values}

\item{error}{Error for convergence}

\item{lambda}{Regularization parameter controlling the nuclear norm of the minimizer}

\item{silence}{Choice to mute iteration. T: mute; F: track each iteration}
}
\value{
Complete matrix
}
\description{
Low rank matrix completion
}
\details{
This package provide SOFT-IMPUTE algorithm for the nuclear norm regularized least-squares problem that scales to large problems. Iteratively, this algorithm replaces the missing elements with those obtained from a soft-thresholded SVD. At every iteration SOFT-IMPUTE decreases the value of the objective function towards its minimum, and at the same time gets closer to the set of optimal solutions of the problem. Reference: Rahul Mazumder, Trevor Hastie, and Robert Tibshirani, 2010, Journal of Machine Learning Research 11(2010)2287-2322.
}
\examples{
AdmmMC(x=incompleteMatrix, iniMatrix = NULL, error = 1e-16, lambda = 0.5, silence = F)

}
\author{
Ye Yuan, Qianyu Wen
}
